Extract wave source analytics aggregation
